revert(admin): remove temporary search canary auth test
