+++ apps/admin/src/app/watch/demo-keyword-search/demo-search-client.tsx
@@ -2,8 +2,13 @@
 
 import { useRouter, useSearchParams } from "next/navigation"
 import { useEffect, useMemo, useState, type FormEvent } from "react"
-import { type SearchResponse, type SearchResult } from "./search-operation"
-import { searchAdminGraphQL } from "./search-action"
+import { executeGraphQL } from "./graphql-client"
+import {
+  SEARCH_OPERATION,
+  type DemoSearchData,
+  type SearchResponse,
+  type SearchResult,
+} from "./search-operation"
 import {
   buildProvenanceMap,
   computeThreeWayDiff,
@@ -208,7 +213,7 @@ export function DemoSearchClient() {
         onSubmit={handleSubmit}
         style={{
           display: "grid",
-          gridTemplateColumns: "minmax(220px, 1fr) 120px 100px 100px auto",
+          gridTemplateColumns: "1fr 120px 100px 100px auto",
           gap: 8,
           alignItems: "end",
           marginBottom: 16,
@@ -300,7 +305,14 @@ async function runSearch(args: {
   limit: number
   mode: ModeKey
 }): Promise<SearchResponse> {
-  return await searchAdminGraphQL(args)
+  const result = await executeGraphQL<DemoSearchData, typeof args>(
+    SEARCH_OPERATION,
+    args,
+  )
+  if (!result.ok) {
+    throw new Error(result.errors.map((e) => e.message).join("; "))
+  }
+  return result.data.search
 }
 
 function toPaneState(settled: PromiseSettledResult<SearchResponse>): PaneState {
