+++ apps/admin/src/app/watch/demo-keyword-search/graphql-client.ts
@@ -1,8 +1,7 @@
 /**
  * Thin fetch wrapper for admin's own GraphQL endpoint at /api/graphql.
  *
- * Defaults to same-origin browser POST; server actions may pass an
- * absolute endpoint and server-side bearer. Returns a discriminated outcome:
+ * Same-origin POST from the browser. Returns a discriminated outcome:
  * - { ok: true, data } when the response carries data with no errors
  * - { ok: false, errors } when GraphQL returned errors[]
  * - { ok: false, errors: [{ message: ... }] } on network / parse failure
@@ -15,39 +14,16 @@ export type GraphQLResult<T> =
   | { ok: true; data: T }
   | { ok: false; errors: Array<{ message: string }> }
 
-export type ExecuteGraphQLOptions = {
-  bearerToken?: string
-  endpoint?: string
-  origin?: string | null
-}
-
-function authorizationHeaderValue(token: string | undefined): string | null {
-  const trimmed = token?.trim()
-  if (!trimmed) return null
-  return /^Bearer\s+/i.test(trimmed) ? trimmed : `Bearer ${trimmed}`
-}
-
 export async function executeGraphQL<
   TData,
   TVars extends Record<string, unknown>,
->(
-  query: string,
-  variables: TVars,
-  options: ExecuteGraphQLOptions = {},
-): Promise<GraphQLResult<TData>> {
+>(query: string, variables: TVars): Promise<GraphQLResult<TData>> {
   let response: Response
-  const authorization = authorizationHeaderValue(options.bearerToken)
-  const origin = options.origin?.trim()
   try {
-    response = await fetch(options.endpoint ?? "/api/graphql", {
+    response = await fetch("/api/graphql", {
       method: "POST",
-      headers: {
-        "Content-Type": "application/json",
-        ...(authorization ? { Authorization: authorization } : {}),
-        ...(origin ? { Origin: origin } : {}),
-      },
+      headers: { "Content-Type": "application/json" },
       body: JSON.stringify({ query, variables }),
-      cache: "no-store",
     })
   } catch (error) {
     return {
